fixed warnigs

+++ src/gameboy/emulator.cpp
@@ -40,7 +40,7 @@ bool loadGB(const std::string& file_name, std::vector<uint8_t>& buffer)
     std::ifstream file(file_name, std::ios::binary | std::ios::ate);
     if(file.is_open())
     {
-        size_t length = file.tellg();
+        size_t length = (size_t)file.tellg();
         buffer.resize(length);
 
         file.seekg(0, std::ios::beg);
